Add style in User page.
