+++ package.json
@@ -14,6 +14,7 @@
     "react-hook-form": "^7.34.2",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
+    "sass": "^1.54.8",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
