+++ src/App.js
@@ -5,7 +5,7 @@ import Layout from "./components/layout/Layout";
 import Home from "./pages/Home";
 import ReinitPassword from "./pages/ReinitPassword";
 import Sendemail from "./pages/Sendemail";
-import User from "./pages/User";
+import User from "./pages/user/User";
 import ArticleDetails from "./pages/ArticleDetails";
 
 function App() {
