+++ src/components/article/article.scss
@@ -1,0 +1,11 @@
+.card-container{
+    display:grid;
+    grid-template-columns: 1fr;
+    justify-items: center;
+
+}
+
+img{
+    width: 80%;
+    height: auto;
+}
