+++ src/pages/Home.js
@@ -2,7 +2,7 @@ import { useAtomValue, useSetAtom } from "jotai";
 import React, { useEffect, useReducer, useState } from "react";
 import { logged } from "../components/atoms/logged";
 import { API } from "../utils/variables";
-import Article from "../components/Article";
+import Article from "../components/article/Article";
 import CreateArticle from "../components/CreateArticle";
 
 function Home(props) {
