+++ src/pages/user/user.scss
@@ -1,0 +1,26 @@
+#grid-container{
+    // border: 3px solid red;
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr;
+    justify-items: center;
+    align-items: center;
+}
+
+#form-container{
+    border: 1px solid #5c76a0;
+    border-radius: 20px;
+    padding: 5px;
+    width: 25vw;
+    box-shadow: 0px 0px 10px 1px #0f151e ;
+    align-self: start;
+
+    h1{
+        background-color: #1e293b;
+        color: white;
+        font-size: 1.5rem;
+        font-weight: 600;
+        border-radius: 20px 20px 0 0;
+        padding: 5px;
+        text-align: center;
+    }
+}
